handle write errors
adds an error argument to the chunk write request callback, so the
creator of the cwr gets to know whether the write was a success or not.
in the case of the mt-whisper-importer-writer this is used to return an
error to the mt-whisper-importer-reader in case there was an error,
which can then retry to submit its request.